+++ src/constants.cpp
@@ -3,6 +3,6 @@
 namespace constants {
 
 const int g_decimalBase = 10;
-const int g_oneMegabyte = 1000000;
+const size_t g_oneMegabyte = 1000000;
 
 } // constants
